Refactor SCTP constants and improve data handling in RTCDtlsTransport and InboundStream
