+++ packages/webrtc/tests/transport/dtls.test.ts
@@ -6,7 +6,7 @@ describe("RTCDtlsTransportTest", () => {
   test("dtls_test_data", async () => {
     const [session1, session2] = await dtlsTransportPair();
     const receiver2 = new DummyDataReceiver();
-    session2.dataReceiver = receiver2.handleData;
+    session2.onData.subscribe((data) => receiver2.handleData(data));
 
     session1.sendData(Buffer.from("ping"));
     await setTimeout(100);
